+++ 2.7.x/choicemaker-cm/choicemaker-swing/com.choicemaker.cm.modelmaker/src/com/choicemaker/cm/modelmaker/gui/dialogs/CollectionsDialog.java
@@ -439,7 +439,7 @@ public class CollectionsDialog extends JDialog implements Enable {
 			int visibleHeight = (int) d.getHeight();
 			
 			if (scrollY + visibleHeight > viewHeight) {
-				scrollY = (int) Math.max(0, viewHeight - visibleHeight);	
+				scrollY = Math.max(0, viewHeight - visibleHeight);	
 			}
 			
 			// move the view
